added InitializeScreen to conditionally navigate to AuthScreen or MainTabs; renamed navigation.js to navigations.js
