+++ src/screens/Auth/Auth.js
@@ -9,7 +9,7 @@ import DefaultInput from '../../components/UI/DefaultInput/DefaultInput'
 import MainText from '../../components/UI/MainText/MainText'
 import TextButton from '../../components/UI/TextButton/TextButton'
 // import ButtonWithBackground from '../../components/UI/ButtonWithBackground/ButtonWithBackground'
-// import { startMainTabs } from './src/screens/navigation'
+// import { startMainTabs } from './src/screens/navigations'
 import validate from '../../utility/validation'
 import { signUpAction, logInAction } from '../../store/actions/auth'
 // import { FACEBOOK, GOOGLE } from '../../constants'
